use once on retriever, refactor to router, add api
