+++ retriever.js
@@ -180,7 +180,7 @@ function genStats() {
 
 function getMMStats(idx, cb) {
     steamObj[idx].Dota2.matchmakingStatsRequest();
-    steamObj[idx].Dota2.on('matchmakingStatsData', function(waitTimes, searchingPlayers, disabledGroups, raw){
+    steamObj[idx].Dota2.once('matchmakingStatsData', function(waitTimes, searchingPlayers, disabledGroups, raw){
         cb(null, raw.searching_players_by_group_source2);
     });
 }
